requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
flake8==3.8.4
mccabe==0.6.1
mysqlclient==2.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.4.1
